refactor: move FormalFifoCCTester
